fixed a shit ton of errors from not using comp flags lol

+++ inc/vec_lib.h
@@ -13,6 +13,8 @@
 #ifndef VEC_LIB_H
 # define VEC_LIB_H
 
+# include <stddef.h>
+
 # define TOKEN_INIT(el, size) element el;\
     element_init(&el, size)
 
@@ -34,9 +36,9 @@ struct s_element
     int (*pf_element_add)(element *, void *);
     int (*pf_element_swap)(element *, void *, size_t pos1, size_t pos2);
     int (*pf_element_insert)(element *, void *, int pos);
-    int (*pf_element_set)(element *, int, void *);
-    void *(*pf_element_get)(element *, int);
-    int (*pf_element_delete)(element *, int);
+    int (*pf_element_set)(element *, size_t, void *);
+    void *(*pf_element_get)(element *, size_t);
+    int (*pf_element_delete)(element *, size_t);
     int (*pf_element_free)(element *);
 };
 
@@ -45,9 +47,9 @@ int element_resize(element *e, int oldsize, int newsize);
 int element_push_back(element *e, void *token);
 int element_swap(element *e, void *token, size_t pos1, size_t pos2);
 int element_insert(element *e, void *token, int pos);
-int element_set(element *e, int index, void *token);
-void *element_get(element *e, int index);
-int element_delete(element *e, int index);
+int	element_set(element *e, size_t index, void *token);
+void	*element_get(element *e, size_t index);
+int element_delete(element *e, size_t index);
 int element_free(element *e);
 void element_init(element *e, int size);
